Seperate usermod into createUser and editUser scripts
