+++ INSTALL.md
@@ -37,7 +37,7 @@ OED_LOG_FILE=?                 // Path to the log file, defaults to ./log.txt
 8. Run ```npm run createdb``` to create the database schema.
 1. Run `npm run addMamacMeters` to load mamac meters from an `.csv` file.
 1. Run `npm run updateMamacMeters` to fetch new data for mamac meters in the database.
-1. Run `npm run usermod` and follow the directions to create a new admin user.
+1. Run `npm run createUser` and follow the directions to create a new admin user.
 1. Run ```npm run build``` to create the Webpack bundle for production, otherwise run ```npm run dev``` for development.
 1. Run ```npm start```
 
@@ -66,6 +66,7 @@ You can access the PostgreSQL database through the `database` service. Given tha
 * Get a Postgres shell with `docker-compose exec database psql -U oed`
 * Take a database dump with `docker-compose exec database pg_dump -U oed > dump_$(date +%Y-%m-%d"_"%H_%M_%S.sql)`
 * Restore a database dump with by first copying the dump into the container with `docker cp /path/to/dump.sql container_name:/dump.sql` and then restoring it into the database with `docker-compose exec database psql -U oed -f /dump.sql`.
+* Change an admin's password with `docker-compose run --rm web npm run editUser` and follow the directions.
 
 ### Upgrading the App ###
